move item-related combat options to item config
