+++ src/main/java/com/github/thedeathlycow/scorchful/config/section/CombatConfig.java
@@ -27,22 +27,10 @@ public class CombatConfig {
     private static final String CATEGORY = ScorchfulConfig.MAIN_CATEGORY_NAME;
 
     @AutoGen(category = CATEGORY)
-    @Translate.Name("Fireball throwing type")
-    @SerialEntry(comment = "Controls what type of fireball is creating when throwing a fire charge. Small is just like firing from a Dispener and creates a fire on impact. Large is more like a Ghast and creates an explosion, and disabled disables this feature.")
-    @EnumCycler
-    FireChargeThrower.FireballFactory fireBallThrownType = FireChargeThrower.FireballFactory.SMALL;
-
-    @AutoGen(category = CATEGORY)
     @Translate.Name("Enable desert visions")
     @SerialEntry(comment = "Toggles hallucinations when overheating in the desert like boats and flowers.")
     @TickBox
     boolean enableDesertVisions = true;
-
-    @AutoGen(category = CATEGORY)
-    @Translate.Name("Fire Protection heat resistance per level")
-    @SerialEntry(comment = "How much Heat Resistance the Fire Protection enchantment should give, per level of Fire Protection.")
-    @DoubleField
-    double fireProtectionHeatResistancePerLevel = 0.125;
 
     @AutoGen(category = CATEGORY)
     @Translate.Name("Fear detection range multiplier")
@@ -50,29 +38,11 @@ public class CombatConfig {
     @DoubleField(min = 0, max = 128)
     double fearDetectionRangeMultiplier = 2.0;
 
-    @AutoGen(category = CATEGORY)
-    @Translate.Name("Impaling damage per level")
-    @SerialEntry(comment = "How much damage the Impaling enchantment should do to wet entities, per level of Impaling")
-    @FloatField(min = 0f)
-    float impalingDamagePerLevel = 2.5f;
-
-    public FireChargeThrower.FireballFactory getFireBallThrownType() {
-        return fireBallThrownType;
-    }
-
     public boolean enableDesertVisions() {
         return enableDesertVisions;
     }
 
-    public double getFireProtectionHeatResistancePerLevel() {
-        return fireProtectionHeatResistancePerLevel;
-    }
-
     public double getFearDetectionRangeMultiplier() {
         return fearDetectionRangeMultiplier;
-    }
-
-    public float getImpalingDamagePerLevel() {
-        return impalingDamagePerLevel;
     }
 }
